+++ src/Repositories/AbstractJsonApiDoctrineRepository.php
@@ -153,7 +153,7 @@ abstract class AbstractJsonApiDoctrineRepository
      */
     public function save(object $entity, bool $mute = false): void
     {
-        $this->isValidEntity($entity);
+        $this->assertValidEntity($entity);
         $this->em->persist($entity);
 
         if (!$mute) {
@@ -168,7 +168,7 @@ abstract class AbstractJsonApiDoctrineRepository
      */
     public function delete(object $entity, bool $mute = false): void
     {
-        $this->isValidEntity($entity);
+        $this->assertValidEntity($entity);
         $this->em->remove($entity);
 
         if (!$mute) {
@@ -198,7 +198,7 @@ abstract class AbstractJsonApiDoctrineRepository
      *
      * @phpstan-param  Entity $entity
      */
-    protected function isValidEntity(object $entity): void
+    protected function assertValidEntity(object $entity): void
     {
         if (!is_a($entity, $this->class)) {
             throw new \Exception('Invalid Entity: ' . get_class($entity));