+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2022-10-21
+
+### Fixed
+
+-   Correct return type for `FilterManager@getValidKey()` to support an array of
+    values for `key` field in ID and Keyword filters.
+
 ## [4.0.1] - 2022-10-04
 
 ### Fixed