+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.0] - 2023-01-18 
+
+### Changed
+
+-   Remove deprecated Doctrine methods.
+-   Changed `isValidEntity` to `assertValidEntity`.
+
 ## [4.1.0] - 2022-11-10 
 
 ### Added